LayoutTests/fast/dom/Document/scrollingElement-standards-mode.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
if (window.internals)
    internals.settings.setCSSOMViewScrollingAPIEnabled(true);

description("Tests the behavior of document.scrollingElement in standards mode.");

debug("In standards mode the scrolling element is supposed to be `HTML`");
shouldBe("document.scrollingElement", "document.documentElement");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
